Implemented forgot password functionality
Now user can click on forgot password and is redirected to another screen. there he can enter his email to which a reset token will be sent. He is then redirected to another page where the user can enter the code and specify a new password. If successfully he will be redirected to the login page again.